Update to 1.21.5
This commit is a squash of all commits from the "Update minecraft 1.21" PR on the original Valence report combined with additional final fixes applied later. Please check the links below for the details:

https://github.com/valence-rs/valence/pull/675

https://github.com/ChunkEdge/ChunkEdge/commit/5b2290a7f0e1f7d15cd4bf309a8ce8942511fb11
https://github.com/ChunkEdge/ChunkEdge/pull/7
https://github.com/ChunkEdge/ChunkEdge/pull/8
https://github.com/ChunkEdge/ChunkEdge/pull/9
https://github.com/ChunkEdge/ChunkEdge/pull/10
https://github.com/ChunkEdge/ChunkEdge/pull/11
https://github.com/ChunkEdge/ChunkEdge/pull/12

+++ crates/valence_entity/src/active_status_effects.rs
@@ -307,9 +307,10 @@ pub struct ActiveStatusEffect {
     effect: StatusEffect,
     /// # Default Value
     /// 0
-    amplifier: u8,
+    amplifier: i32,
     /// The initial duration of the effect in ticks.
     /// If `None`, the effect is infinite.
+    /// (encoded as -1 if infinite)
     ///
     /// # Default Value
     /// Some(600) (30 seconds)
@@ -345,7 +346,7 @@ impl ActiveStatusEffect {
     }
 
     /// Sets the amplifier of the [`ActiveStatusEffect`].
-    pub fn with_amplifier(mut self, amplifier: u8) -> Self {
+    pub fn with_amplifier(mut self, amplifier: i32) -> Self {
         self.amplifier = amplifier;
         self
     }
@@ -397,7 +398,7 @@ impl ActiveStatusEffect {
     }
 
     /// Returns the amplifier of the [`ActiveStatusEffect`].
-    pub fn amplifier(&self) -> u8 {
+    pub fn amplifier(&self) -> i32 {
         self.amplifier
     }
